@@ -16,8 +16,7 @@
 
 package org.springframework.guice;
 
-import javax.inject.Inject;
-
+import jakarta.inject.Inject;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 
