@@ -16,13 +16,12 @@
 
 package org.springframework.guice;
 
-import javax.inject.Singleton;
-
 import com.google.inject.AbstractModule;
 import com.google.inject.Guice;
 import com.google.inject.Injector;
 import com.google.inject.TypeLiteral;
 import com.google.inject.name.Names;
+import jakarta.inject.Singleton;
 
 /**
  * @author Dave Syer
