@@ -16,13 +16,12 @@
 
 package org.springframework.guice.annotation;
 
-import javax.inject.Inject;
-import javax.inject.Named;
-
 import com.google.inject.AbstractModule;
 import com.google.inject.Injector;
 import com.google.inject.Provides;
 import com.google.inject.Singleton;
+import jakarta.inject.Inject;
+import jakarta.inject.Named;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 
